SLA-3638 Refactoring

+++ Bundles/ImportProcessGoogleSheets/src/SprykerDemo/Zed/ImportProcessGoogleSheets/Business/Payload/Deleter/ImportProcessPayloadCsvDataDeleter.php
@@ -21,11 +21,7 @@ class ImportProcessPayloadCsvDataDeleter implements ImportProcessPayloadDataDele
         foreach ($importProcessTransfer->getPayloadOrFail()->getSourceMaps() as $sourceMapTransfer) {
             $filePath = $sourceMapTransfer->getSource();
 
-            if ($filePath === null) {
-                return $importProcessTransfer;
-            }
-
-            if (file_exists($filePath)) {
+            if ($filePath !== null && file_exists($filePath)) {
                 unlink($filePath);
             }
         }
